# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ClientAccountInput {
  /**
     * @minLength 1
     * @maxLength 120
     */
  fullName: string;
  email: string;
  /**
     * @minLength 8
     * @maxLength 128
     */
  initialPassword: string;
}

export type ClientAccountRole = typeof ClientAccountRole[keyof typeof ClientAccountRole];


export const ClientAccountRole = {
  client: 'client',
} as const;

export type ClientAccountStatus = typeof ClientAccountStatus[keyof typeof ClientAccountStatus];


export const ClientAccountStatus = {
  active: 'active',
  unconfirmed: 'unconfirmed',
  disabled: 'disabled',
} as const;

export interface ClientAccount {
  id: string;
  fullName: string;
  email: string;
  role: ClientAccountRole;
  status: ClientAccountStatus;
}