# lib/api-zod/src/generated/types/clientAccount.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ClientAccountRole } from './clientAccountRole';
import type { ClientAccountStatus } from './clientAccountStatus';

export interface ClientAccount {
  id: string;
  fullName: string;
  email: string;
  role: ClientAccountRole;
  status: ClientAccountStatus;
}