# lib/api-zod/src/generated/types/clientAccountInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ClientAccountInput {
  /**
     * @minLength 1
     * @maxLength 120
     */
  fullName: string;
  email: string;
  /**
     * @minLength 8
     * @maxLength 128
     */
  initialPassword: string;
}